+++ CHANGELOG.md
@@ -2,5 +2,7 @@
 
 | Version | Date     | Changes                   |
 | ------- | -------- | ------------------------- |
-|2.1.0    |2022-12-15| - added Utils for Coding, HumanName and Period<br />- added tests for utils|
+|2.1.2    |2022-12-21| - Add getIdentifierString() to utils |
+|2.1.1    |2022-12-16| - Fix a bug in bundle smart resource |
+|2.1.0    |2022-12-15| - added Utils for Coding, HumanName and Period<br />- added tests for utils |
 |2.0.0    |2022-11   | predates Changelog        |